+++ bastion/resource_connection_policy.go
@@ -59,7 +59,7 @@ func resourceConnectionPolicy() *schema.Resource {
 	}
 }
 func resourceConnectionPolicyVersionCheck(version string) error {
-	if version == versionValidate3_3 {
+	if bchk.StringInSlice(version, defaultVersionsValid()) {
 		return nil
 	}