+++ bastion/resource_device_localdomain_account.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	bchk "github.com/jeremmfr/go-utils/basiccheck"
 )
 
 type jsonDeviceLocalDomainAccount struct {
@@ -107,7 +108,7 @@ func resourceDeviceLocalDomainAccount() *schema.Resource {
 	}
 }
 func resourceDeviceLocalDomainAccountVersionCheck(version string) error {
-	if version == versionValidate3_3 {
+	if bchk.StringInSlice(version, defaultVersionsValid()) {
 		return nil
 	}