+++ bastion/resource_externalauth_radius.go
@@ -9,6 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	bchk "github.com/jeremmfr/go-utils/basiccheck"
 )
 
 type jsonExternalAuthRadius struct {
@@ -67,7 +68,7 @@ func resourceExternalAuthRadius() *schema.Resource {
 	}
 }
 func resourceExternalAuthRadiusVersionCheck(version string) error {
-	if version == versionValidate3_3 {
+	if bchk.StringInSlice(version, defaultVersionsValid()) {
 		return nil
 	}
 
@@ -275,9 +276,6 @@ func fillExternalAuthRadius(d *schema.ResourceData, jsonData jsonExternalAuthRad
 		panic(tfErr)
 	}
 	if tfErr := d.Set("port", jsonData.Port); tfErr != nil {
-		panic(tfErr)
-	}
-	if tfErr := d.Set("secret", jsonData.Secret); tfErr != nil {
 		panic(tfErr)
 	}
 	if tfErr := d.Set("timeout", jsonData.Timeout); tfErr != nil {