+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 <!-- markdownlint-disable-file MD013 MD041 -->
 # changelog
 
+FEATURES:
+
+* add `wallix-bastion_version` data source
+
+ENHANCEMENTS:
+
+* allow use `v3.6` to `api_version` provider argument
+
+BUG FIXES:
+
+* resource/**wallix-bastion_externalauth_kerberos**: add `keytab` argument required in latest version of WAB
+* resource/**wallix-bastion_externalauth_radius**: `secret` argument can't be refresh in latest version of WAB
+* resource/**wallix-bastion_externalauth_tacacs**: `secret` argument can't be refresh in latest version of WAB
+
 ## 0.5.0 (December 9, 2021)
 
 NOTES: